requirements.txt
blinker==1.4
flask-login==0.3.2
flask-oauthlib==0.9.2
flask-restful==0.3.5
flask-wtf==0.12
flask==0.10.1
GoogleAppEngineCloudStorageClient==1.9.22.1
unidecode==0.4.18
webargs==1.2.0
